# api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NFC Loyalty & Customer Growth Platform API",
    "description": "REST API untuk platform loyalty berbasis NFC/QR untuk UMKM Indonesia. Menyediakan fitur registrasi member, pemesanan, validasi pembayaran manual, poin loyalitas, voucher, dan dashboard analytics.",
    "version": "1.0.0",
    "contact": { "name": "NFC Loyalty Platform" }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Development server" }
  ],
...
  "paths": {
...
    "/api/orders/checkout": {
      "post": {
        "tags": ["Orders"],
        "summary": "Checkout order (member)",
        "description": "Create new order from cart items. Optionally apply voucher code.",
        "operationId": "checkoutOrder",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CheckoutRequest" },
              "example": {
                "items": [
                  { "menuItemId": "e5f6a7b8-c9d0-1234-efgh-567890123456", "quantity": 2 },
                  { "menuItemId": "f6a7b8c9-d012-3456-fghi-678901234567", "quantity": 1 }
                ],
                "voucherCode": "DISKON10"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "orderId": "f7a8b9c0-d1e2-3456-ghij-789012345678",
                    "originalTotal": 70000,
                    "discountAmount": 7000,
                    "finalTotal": 63000,
                    "paymentBarcode": "ORD-f7a8b9c0-63000",
                    "status": "pending"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid voucher",
            "content": {
              "application/json": {
                "example": { "success": false, "error": { "message": "Voucher sudah expired" } }
              }
            }
          }
        }
      }
    },
    "/api/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get order history (member)",
        "operationId": "getMemberOrders",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated order history",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "orders": [
                      {
                        "id": "f7a8b9c0-d1e2-3456-ghij-789012345678",
                        "originalTotal": 70000,
                        "discountAmount": 7000,
                        "finalTotal": 63000,
                        "status": "paid",
                        "pointsEarned": 6,
                        "paymentBarcode": "ORD-f7a8b9c0-63000",
                        "createdAt": "2025-07-14T12:00:00.000Z",
                        "validatedAt": "2025-07-14T12:15:00.000Z",
                        "items": [
                          { "id": "oi-001", "itemName": "Nasi Goreng Spesial", "itemPrice": 25000, "quantity": 2 },
                          { "id": "oi-002", "itemName": "Es Teh Manis", "itemPrice": 8000, "quantity": 1 }
                        ]
                      }
                    ],
                    "pagination": { "page": 1, "pageSize": 20, "total": 8, "totalPages": 1 }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "List orders (admin)",
        "description": "Filterable by status. Pending sorted oldest first, others newest first.",
        "operationId": "listAdminOrders",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending", "paid", "cancelled", "expired"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 50, "maximum": 100 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated order list",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "orders": [
                      {
                        "id": "f7a8b9c0-d1e2-3456-ghij-789012345678",
                        "originalTotal": 70000,
                        "discountAmount": 7000,
                        "finalTotal": 63000,
                        "status": "pending",
                        "paymentBarcode": "ORD-f7a8b9c0-63000",
                        "createdAt": "2025-07-14T12:00:00.000Z",
                        "member": { "id": "a1b2c3d4", "name": "John Doe", "whatsapp": "081234567890" },
                        "items": [
                          { "itemName": "Nasi Goreng Spesial", "itemPrice": 25000, "quantity": 2 }
                        ],
                        "voucher": { "code": "DISKON10", "discountType": "percentage", "discountValue": 10 }
                      }
                    ],
                    "pagination": { "page": 1, "pageSize": 50, "total": 3, "totalPages": 1 }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/orders/{orderId}/validate": {
      "post": {
        "tags": ["Orders"],
        "summary": "Validate payment (admin confirm/reject)",
        "operationId": "validatePayment",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "orderId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["action"],
                "properties": { "action": { "type": "string", "enum": ["confirm", "reject"] } }
              },
              "example": { "action": "confirm" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment validated",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "orderId": "f7a8b9c0-d1e2-3456-ghij-789012345678",
                    "status": "paid",
                    "pointsAwarded": 6,
                    "notificationSent": true
                  }
                }
              }
            }
          },
          "400": {
            "description": "Order already in terminal state",
            "content": {
              "application/json": {
                "example": { "success": false, "error": { "message": "Order sudah dalam status paid, tidak bisa diubah" } }
              }
            }
          }
        }
      }
    },
...
    "/api/rewards": {
      "get": {
        "tags": ["Rewards & Points"],
        "summary": "List redeemable rewards (member)",
        "description": "Returns rewards that are active and affordable for the member.",
        "operationId": "getRedeemableRewards",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of redeemable rewards",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "rewards": [
                      {
                        "id": "r-001",
                        "name": "Free Coffee",
                        "description": "Gratis 1 kopi espresso",
                        "requiredPoints": 50,
                        "stockQuantity": 20,
                        "isActive": true
                      },
                      {
                        "id": "r-002",
                        "name": "Discount 50%",
                        "description": "Potongan 50% untuk 1 menu",
                        "requiredPoints": 100,
                        "stockQuantity": 5,
                        "isActive": true
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/rewards/{id}/redeem": {
      "post": {
        "tags": ["Rewards & Points"],
        "summary": "Redeem a reward (member)",
        "operationId": "redeemReward",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Reward redeemed successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "rewardName": "Free Coffee",
                    "pointsDeducted": 50,
                    "remainingBalance": 100
                  }
                }
              }
            }
          },
          "400": {
            "description": "Cannot redeem",
            "content": {
              "application/json": {
                "examples": {
                  "insufficient": { "summary": "Not enough points", "value": { "success": false, "error": { "message": "Poin tidak cukup. Dibutuhkan: 50, saldo: 30" } } },
                  "inactive": { "summary": "Reward unavailable", "value": { "success": false, "error": { "message": "Reward tidak tersedia" } } }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/rewards": {
      "post": {
        "tags": ["Rewards & Points"],
        "summary": "Create reward (admin)",
        "operationId": "createReward",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateRewardRequest" },
              "example": {
                "name": "Free Coffee",
                "description": "Gratis 1 kopi espresso",
                "requiredPoints": 50,
                "stockQuantity": 20,
                "isActive": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reward created",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "reward": {
                      "id": "r-003",
                      "tenantId": "550e8400-e29b-41d4-a716-446655440000",
                      "name": "Free Coffee",
                      "description": "Gratis 1 kopi espresso",
                      "requiredPoints": 50,
                      "stockQuantity": 20,
                      "isActive": true,
                      "createdAt": "2025-07-15T10:30:00.000Z"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/rewards/{id}": {
      "patch": {
        "tags": ["Rewards & Points"],
        "summary": "Update reward (admin)",
        "operationId": "updateReward",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateRewardRequest" },
              "example": {
                "name": "Free Latte",
                "requiredPoints": 75,
                "stockQuantity": 10,
                "isActive": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reward updated",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "reward": {
                      "id": "r-001",
                      "name": "Free Latte",
                      "requiredPoints": 75,
                      "stockQuantity": 10,
                      "isActive": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/members/me/points": {
      "get": {
        "tags": ["Rewards & Points"],
        "summary": "Get point transaction history (member)",
        "operationId": "getPointHistory",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Point transaction history",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "transactions": [
                      {
                        "id": "pt-001",
                        "memberId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                        "type": "earned",
                        "amount": 6,
                        "orderId": "f7a8b9c0-d1e2-3456-ghij-789012345678",
                        "rewardId": null,
                        "resultingBalance": 156,
                        "createdAt": "2025-07-14T12:15:00.000Z"
                      },
                      {
                        "id": "pt-002",
                        "memberId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                        "type": "redeemed",
                        "amount": 50,
                        "orderId": null,
                        "rewardId": "r-001",
                        "resultingBalance": 150,
                        "createdAt": "2025-07-13T09:00:00.000Z"
                      }
                    ],
                    "pagination": { "page": 1, "pageSize": 20, "total": 25, "totalPages": 2 }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token (15min). Get from /api/auth/register, /api/auth/login, or /api/auth/admin/login"
      }
    },
    "schemas": {
...
      "CheckoutRequest": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "object",
              "required": ["menuItemId", "quantity"],
              "properties": {
                "menuItemId": { "type": "string", "format": "uuid" },
                "quantity": { "type": "integer", "minimum": 1, "maximum": 99 }
              }
            }
          },
          "voucherCode": { "type": "string", "maxLength": 20 }
        }
      },
...
      "CreateRewardRequest": {
        "type": "object",
        "required": ["name", "requiredPoints", "stockQuantity"],
        "properties": {
          "name": { "type": "string", "maxLength": 100 },
          "description": { "type": "string", "maxLength": 500 },
          "requiredPoints": { "type": "integer", "minimum": 1 },
          "stockQuantity": { "type": "integer", "minimum": 0 },
          "isActive": { "type": "boolean", "default": true }
        }
      },
...
    }
  }
}